// src/types.tsx
export type Alpaca = {
  accessories: string;
  backgrounds: string;
  ears: string;
  eyes: string;
  hair: string;
  leg: string;
  mouth: string;
  neck: string;
  nose: string;
};

export type Settings = {
  accessories: string[];
  backgrounds: string[];
  ears: string[];
  eyes: string[];
  hair: string[];
  leg: string[];
  mouth: string[];
  neck: string[];
  nose: string[];
};